+++ src/components/Footer/Footer.styles.ts
@@ -2,23 +2,15 @@ import { createStyles } from '@mantine/core'
 
 export const useStyles = createStyles(theme => ({
 	footer: {
-		gridArea: 'footer',
-		marginTop: 120,
+		display: 'flex',
+		justifyContent: 'space-around',
+		alignItems: 'center',
+		padding: 10,
+		maxHeight: 60,
 		borderTop: `1px solid ${
 			theme.colorScheme === 'dark' ? theme.colors.dark[5] : theme.colors.gray[2]
 		}`,
 		background: theme.colors.dark[7]
-	},
-
-	inner: {
-		display: 'flex',
-		justifyContent: 'space-between',
-		alignItems: 'center',
-		padding: `${theme.spacing.md}px ${theme.spacing.md}px`,
-
-		[theme.fn.smallerThan('sm')]: {
-			flexDirection: 'column'
-		}
 	},
 
 	links: {