+++ src/components/ProjectCard/ProjectCard.styles.ts
@@ -7,7 +7,6 @@ export const useStyles = createStyles(theme => ({
 	},
 
 	section: {
-		cursor: 'default',
 		borderBottom: `1px solid ${
 			theme.colorScheme === 'dark' ? theme.colors.dark[4] : theme.colors.gray[3]
 		}`,