+++ src/pages/index.tsx
@@ -1,5 +1,3 @@
-import { Container } from '@mantine/core'
-
 import { HeroText } from '~/components/Hero/Hero.component'
 
 export default function Home() {