+++ CHANGELOG.md
@@ -2,6 +2,84 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.1.0](https://github.com/HigorAlves/higoralves.dev/compare/v3.0.0...v1.1.0) (2023-04-20)
+
+
+### :rocket: Build System
+
+* fixing build steps ([217da5d](https://github.com/HigorAlves/higoralves.dev/commit/217da5df2e48ec934a014938ab2ea6cdb0870996))
+
+
+### :recycle: Code Refactoring
+
+* merge conflict ([6830ade](https://github.com/HigorAlves/higoralves.dev/commit/6830ade78139c4c63ce6ea9ff2204299a55539e4))
+* right other for my career ([792e27d](https://github.com/HigorAlves/higoralves.dev/commit/792e27df72b30cd7c948d1edb0fd22d1d0b03652))
+* spliting about page into components ([079d7b3](https://github.com/HigorAlves/higoralves.dev/commit/079d7b350120b2ab8276128b655f81099709f132))
+* using new title component at about page ([10b22c2](https://github.com/HigorAlves/higoralves.dev/commit/10b22c21f39d59fff056efb2bc4ad9dc9c3704ef))
+* using new title component at projects page ([f953f30](https://github.com/HigorAlves/higoralves.dev/commit/f953f3046cad05dd4b18753429de5d1b224c9a46))
+* using the new company object ([7afab57](https://github.com/HigorAlves/higoralves.dev/commit/7afab57d02fc7eef445cde42c586314b2b603ccb))
+
+
+### :triangular_flag_on_post: Others
+
+* adding alto qi description texts ([3f46f00](https://github.com/HigorAlves/higoralves.dev/commit/3f46f007e57b255b6c1ef6b76fa5f1c133f5c614))
+* adding basic image carousel system ([ee01395](https://github.com/HigorAlves/higoralves.dev/commit/ee0139528240519e67b348225bbc437c0c2b520a))
+* adding basic skills into projects ([c94c4bb](https://github.com/HigorAlves/higoralves.dev/commit/c94c4bbaef338ba8b15c6099d26b594e37fa7cf7))
+* adding cie digital text ([9a50a97](https://github.com/HigorAlves/higoralves.dev/commit/9a50a9753db2daa529184c4adf7caadd2876f6dd))
+* adding description texts ([d72326f](https://github.com/HigorAlves/higoralves.dev/commit/d72326f5d81578dffad4266ef3d976addb958634))
+* adding edu livre description texts ([bf8918d](https://github.com/HigorAlves/higoralves.dev/commit/bf8918da7cd906c275b51b3aa401546aef7984b5))
+* adding elderly-health-care description texts ([8c4c065](https://github.com/HigorAlves/higoralves.dev/commit/8c4c0657e996b904d7740494fa78faae22c35ce5))
+* adding ivy cash texts ([b5a1782](https://github.com/HigorAlves/higoralves.dev/commit/b5a1782ac8d4a52e7e4e91af38671ed227cb4d89))
+* adding new techs ([7bd3592](https://github.com/HigorAlves/higoralves.dev/commit/7bd35923464de2beb68c42e14f7d1d10d1da90ae))
+* adding project development details ([aec7537](https://github.com/HigorAlves/higoralves.dev/commit/aec753701d1f5a50f7471c4c76e915a8a8f32387))
+* adding resale description texts ([795238b](https://github.com/HigorAlves/higoralves.dev/commit/795238b1de43e126c55e4830e11ed846f1a38f1b))
+* adding sherlock description texts ([e409a0e](https://github.com/HigorAlves/higoralves.dev/commit/e409a0eabd029855f52252663198de2b9b8204a2))
+* adding take description texts ([1ecc432](https://github.com/HigorAlves/higoralves.dev/commit/1ecc4329f8ada62893b1436affb6d27e69b127ce))
+* adding tech into projects ([2c0e649](https://github.com/HigorAlves/higoralves.dev/commit/2c0e649b327926165489b72adec0f609a9b8ce34))
+* adding telus description texts ([90d3f9d](https://github.com/HigorAlves/higoralves.dev/commit/90d3f9db5016958529979ee210f3f411fc9e94ed))
+* adding toro investimentos description texts ([273de72](https://github.com/HigorAlves/higoralves.dev/commit/273de727421ac76a93317c94ac096620bddc0208))
+* adding traders club description texts ([65c2601](https://github.com/HigorAlves/higoralves.dev/commit/65c26013a99a1f2cc2354b1ee93bb3ef40a2dc7d))
+* adding unesco description texts ([a06e428](https://github.com/HigorAlves/higoralves.dev/commit/a06e428e5bbe04e7b0d992843c746e9a39e44081))
+* adding videopeel description texts ([4829c86](https://github.com/HigorAlves/higoralves.dev/commit/4829c86fe67ca1a34bea15d6cdf125fdb6b820ef))
+* adding xqh platform texts ([32ac1fe](https://github.com/HigorAlves/higoralves.dev/commit/32ac1fecae9a182b36919b652e83d4bb6ca13230))
+* adding yarn.lock ([c13efff](https://github.com/HigorAlves/higoralves.dev/commit/c13efff806384664de3aca3eba2db75acbd1e6da))
+* basic skills map ([050823b](https://github.com/HigorAlves/higoralves.dev/commit/050823be2716ec1e866a5791417b6b72c8fadd8f))
+* code quality tools ([#106](https://github.com/HigorAlves/higoralves.dev/issues/106)) ([909de3f](https://github.com/HigorAlves/higoralves.dev/commit/909de3f1ba863dfab799a88dd95598b90299a721))
+* fixing the company list ([f00f917](https://github.com/HigorAlves/higoralves.dev/commit/f00f9179ac1fde171f0c6bbfda797595f58476ed))
+* image update for companies and projects ([de4a61b](https://github.com/HigorAlves/higoralves.dev/commit/de4a61bc4f04ad906f932af77a93d4be20ebd685))
+* initial page with basic carousel ([0aa0b16](https://github.com/HigorAlves/higoralves.dev/commit/0aa0b1687137a38bed1aacdc126d7feb4be14b57))
+* initial skills set ([ee0aed9](https://github.com/HigorAlves/higoralves.dev/commit/ee0aed94238a941dcc1526389d636caebc38a665))
+* new about for projects ([d1987d6](https://github.com/HigorAlves/higoralves.dev/commit/d1987d6439d9afa83d7f3420259993276f109a81))
+* new files for new version ([cd56282](https://github.com/HigorAlves/higoralves.dev/commit/cd562825da08404e235178194be3c1d817df1539))
+* new icons ([3643c0b](https://github.com/HigorAlves/higoralves.dev/commit/3643c0b6da6c0e162fe47acd13a631962f0b486f))
+* new project text ([4edd179](https://github.com/HigorAlves/higoralves.dev/commit/4edd179be7f26cf26f4623eb7caafe129096c9cb))
+* new sections ([6dccacd](https://github.com/HigorAlves/higoralves.dev/commit/6dccacd9fe0859ad7fb6b36af41af496bb661713))
+* page title component ([999af3e](https://github.com/HigorAlves/higoralves.dev/commit/999af3eec897a4543b8b997e2a7bd0978df3643b))
+* **release:** 1.0.0 ([01c5994](https://github.com/HigorAlves/higoralves.dev/commit/01c5994e9ed3861bd93a0f5eb33e35a8a6ccce9a))
+* removing .yarn ([3190ce0](https://github.com/HigorAlves/higoralves.dev/commit/3190ce04fbc8804d7ec57bd0f4dddaaeb368dfde))
+* tech logos ([5a75e6a](https://github.com/HigorAlves/higoralves.dev/commit/5a75e6a5158cefec923f0f4d4cfd5500f596c205))
+* updating date from end date zipdev ([ac917f9](https://github.com/HigorAlves/higoralves.dev/commit/ac917f91c6b30753de4764e101a0e17925c90965))
+* updating dependencies ([034406c](https://github.com/HigorAlves/higoralves.dev/commit/034406cd354c2b6c011377019a1e6ef1aa9e8bea))
+* upgrading dependencies ([acd496c](https://github.com/HigorAlves/higoralves.dev/commit/acd496c768bb1437d86c28c386be36ec90cc33c7))
+* uses initial page data ([93cda54](https://github.com/HigorAlves/higoralves.dev/commit/93cda5499b7b8851c48f47f1692471b71c24d6f4))
+
+
+### :bug: Bug Fixes
+
+* adding icon fix for build ([36b7b3c](https://github.com/HigorAlves/higoralves.dev/commit/36b7b3c4eece53fe1dbc85d04598497619d5d5a3))
+* footer redirect link bug ([b919789](https://github.com/HigorAlves/higoralves.dev/commit/b9197893344dcd94a94407d40d0429fc913600bb))
+* footer spacing ([45f3f52](https://github.com/HigorAlves/higoralves.dev/commit/45f3f525d1d85ee765ddf70a42800e3eec22afcf))
+* icons inside the about page ([d68e699](https://github.com/HigorAlves/higoralves.dev/commit/d68e6997db0642850baec9078cf30ff81d61d5a1))
+* improving the code quality ([03c11f7](https://github.com/HigorAlves/higoralves.dev/commit/03c11f7c4fe81973954c23764ac55950211b703b))
+* lang for cypress bug ([62524a1](https://github.com/HigorAlves/higoralves.dev/commit/62524a109385876c6288c85df0b874dc72a76829))
+* link color ([c51c756](https://github.com/HigorAlves/higoralves.dev/commit/c51c756984588dc7fde89b41ad1fc0bd225366b7))
+* name size for titles ([e7c7b18](https://github.com/HigorAlves/higoralves.dev/commit/e7c7b187d1626fc5e829d107135e7f682d9ac3ba))
+* page data ([07458cd](https://github.com/HigorAlves/higoralves.dev/commit/07458cdbe8391d1bea993ceea979baff05dc0754))
+* required fields ([b943635](https://github.com/HigorAlves/higoralves.dev/commit/b9436354674abac6afa8a9ef8c111351897b2e63))
+* subtitle size for title component ([766fa40](https://github.com/HigorAlves/higoralves.dev/commit/766fa40b7ebacd4fbec315814a7526d0afd257de))
+* typing ([84feaf0](https://github.com/HigorAlves/higoralves.dev/commit/84feaf00fe8a92a8d2b30cec0981acfc5da93da1))
+* workflow templates ([d73d0d3](https://github.com/HigorAlves/higoralves.dev/commit/d73d0d316184fb32b5dec0d2e34acc6e68c66641))
+
 ## [1.0.0](https://github.com/HigorAlves/higoralves.dev/compare/v3.0.0...v1.0.0) (2023-02-11)
 
 ### :triangular_flag_on_post: Others